requirements.txt
channels==3.0.4
Django==3.2
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
python-binance==1.0.15
async-timeout==4.0.2